+++ app/tests/__init__.py
@@ -1,0 +1,3 @@
+TEST_FILES_RECOVERED_PATH = "./seft_files/recover-"
+
+TEST_FILES_PATH = "./test_files/"