+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 ### Unreleased
+  - Initial version of the SEFT publisher service
+  - Namespace config